requirements.txt
Flask
Pillow
torch
torchvision
Flask-Limiter
gunicorn
opencv-python
numpy
huggingface_hub
keras
tensorflow
pandas